refactor(world): consolidate helper functions in helpers.go
Move reusable helper functions from exit_repo.go to helpers.go:
- ulidsToStrings: convert ULID slice to string slice
- stringsToULIDs: parse string slice to ULID slice (with fieldName for context)
- nullableString: convert empty string to nil
- marshalLockData/unmarshalLockData: JSON serialization for lock data

These functions are generic and may be used by other repositories.
Keep nullableLockType in exit_repo.go as it's exit-specific.
